[Refactor] (test) Move the test spec class into single module '_spec'.
